Merge dev.v1.2.2 branch

+++ lib/service/storage/download.dart
@@ -1,9 +1,12 @@
+import 'dart:convert';
 import 'dart:io' show Directory, File;
 
 import 'package:flutter_ahlib/flutter_ahlib.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:manhuagui_flutter/config.dart';
-import 'package:manhuagui_flutter/page/page/glb_setting.dart';
+import 'package:manhuagui_flutter/model/app_setting.dart';
+import 'package:manhuagui_flutter/model/chapter.dart';
+import 'package:manhuagui_flutter/model/manga.dart';
 import 'package:manhuagui_flutter/service/native/android.dart';
 import 'package:manhuagui_flutter/service/storage/storage.dart';
 
@@ -11,9 +14,9 @@ import 'package:manhuagui_flutter/service/storage/storage.dart';
 // path
 // ====
 
-Future<String> _getDownloadImageDirectoryPath(String url) async {
+Future<String> _getDownloadImageFilePath(String url) async {
   var basename = getTimestampTokenForFilename();
-  var extension = PathUtils.getExtension(url.split('?')[0]);
+  var extension = PathUtils.getExtension(url.split('?')[0]); // include "."
   var filename = '$basename$extension';
   var directoryPath = await lowerThanAndroidR()
       ? await getPublicStorageDirectoryPath() // /storage/emulated/0/Manhuagui/manhuagui_image/IMG_20220917_131013_206.jpg
@@ -36,7 +39,7 @@ Future<String> _getDownloadMangaDirectoryPath([int? mangaId, int? chapterId]) as
 
 Future<String> _getDownloadedChapterPageFilePath({required int mangaId, required int chapterId, required int pageIndex, required String url}) async {
   var basename = (pageIndex + 1).toString().padLeft(4, '0');
-  var extension = PathUtils.getExtension(url.split('?')[0]);
+  var extension = PathUtils.getExtension(url.split('?')[0]); // include "."
   var filename = '$basename$extension';
   return PathUtils.joinPath([await _getDownloadMangaDirectoryPath(mangaId, chapterId), filename]);
 }
@@ -60,13 +63,13 @@ Future<File?> getDownloadedChapterPageFile({required int mangaId, required int c
   }
 }
 
-// ========
-// download
-// ========
+// ==============
+// download image
+// ==============
 
 Future<File?> downloadImageToGallery(String url) async {
   try {
-    var filepath = await _getDownloadImageDirectoryPath(url);
+    var filepath = await _getDownloadImageFilePath(url);
     var f = await downloadFile(
       url: url,
       filepath: filepath,
@@ -78,11 +81,11 @@ Future<File?> downloadImageToGallery(String url) async {
       option: DownloadOption(
         behavior: DownloadBehavior.preferUsingCache,
         conflictHandler: (_) async => DownloadConflictBehavior.addSuffix,
-        headTimeout: GlbSetting.global.dlTimeoutBehavior.determineDuration(
+        headTimeout: AppSetting.instance.other.dlTimeoutBehavior.determineValue(
           normal: Duration(milliseconds: DOWNLOAD_HEAD_TIMEOUT),
           long: Duration(milliseconds: DOWNLOAD_HEAD_LTIMEOUT),
         ),
-        downloadTimeout: GlbSetting.global.dlTimeoutBehavior.determineDuration(
+        downloadTimeout: AppSetting.instance.other.dlTimeoutBehavior.determineValue(
           normal: Duration(milliseconds: DOWNLOAD_IMAGE_TIMEOUT),
           long: Duration(milliseconds: DOWNLOAD_IMAGE_LTIMEOUT),
         ),
@@ -123,6 +126,10 @@ Future<bool> downloadChapterPage({required int mangaId, required int chapterId, 
   }
 }
 
+// =====================
+// download task related
+// =====================
+
 Future<void> createNomediaFile() async {
   // 留到 DownloadMangaQueueTask 再捕获异常
   var nomediaPath = PathUtils.joinPath([await _getDownloadMangaDirectoryPath(), '.nomedia']);
@@ -130,6 +137,65 @@ Future<void> createNomediaFile() async {
   if (!(await nomediaFile.exists())) {
     await nomediaFile.create(recursive: true);
   }
+}
+
+Future<bool> writeMetadataFile({required int mangaId, required int chapterId, required Manga manga, required MangaChapter chapter}) async {
+  try {
+    var metadataPath = PathUtils.joinPath([await _getDownloadMangaDirectoryPath(mangaId, chapterId), 'metadata.json']);
+    var metadataFile = File(metadataPath);
+    if (!(await metadataFile.exists())) {
+      await metadataFile.create(recursive: true);
+    }
+
+    var obj = <String, dynamic>{
+      'version': 1,
+      'manga_id': mangaId,
+      'chapter_id': chapterId,
+      'next_cid': chapter.nextCid,
+      'prev_cid': chapter.prevCid,
+      'pages': chapter.pages,
+    };
+    var encoder = JsonEncoder.withIndent('  ');
+    var content = encoder.convert(obj);
+    await metadataFile.writeAsString(content, flush: true);
+    return true;
+  } catch (e, s) {
+    globalLogger.e('writeMetadataFile', e, s);
+    return false;
+  }
+}
+
+Future<Tuple3<List<String>, int?, int?>> readMetadataFile({required int mangaId, required int chapterId, required int pageCount}) async {
+  List<String>? pages;
+  int? nextCid;
+  int? prevCid;
+  try {
+    var metadataPath = PathUtils.joinPath([await _getDownloadMangaDirectoryPath(mangaId, chapterId), 'metadata.json']);
+    var metadataFile = File(metadataPath);
+    if (await metadataFile.exists()) {
+      var content = await metadataFile.readAsString();
+      var obj = json.decode(content) as Map<String, dynamic>; // version: 1
+      pages = (obj['pages'] as List<dynamic>?)?.map((e) => e.toString()).toList();
+      nextCid = obj['next_cid'] as int?;
+      prevCid = obj['prev_cid'] as int?;
+    } else {
+      globalLogger.w('readMetadataFile "$metadataPath" not found');
+    }
+  } catch (e, s) {
+    globalLogger.e('readMetadataFile', e, s);
+  }
+
+  pages ??= <String>[];
+  if (pageCount > pages.length) {
+    pages.addAll(List.generate(pageCount - pages.length, (i) => '<placeholder_${i + 1}>.webp')); // extension defaults to webp
+  } else if (pages.length > pageCount) {
+    pages = pages.sublist(0, pageCount);
+  }
+  return Tuple3(pages, nextCid, prevCid);
+}
+
+bool isPageUrlValidInMetadata(String url) {
+  return url.isNotEmpty && !url.startsWith('<placeholder_');
 }
 
 Future<int> getDownloadedMangaBytes({required int mangaId}) async {
@@ -173,7 +239,9 @@ Future<bool> deleteDownloadedChapter({required int mangaId, required int chapter
   try {
     var mangaPath = await _getDownloadMangaDirectoryPath(mangaId, chapterId);
     var directory = Directory(mangaPath);
-    await directory.delete(recursive: true);
+    if (await directory.exists()) {
+      await directory.delete(recursive: true);
+    }
     return true;
   } catch (e, s) {
     globalLogger.e('deleteDownloadedChapter', e, s);
